refactor(upgrade): update version check in upgrade controller
Adapt the version-checking part of the upgrade controller to use the
`versionguard` package. To avoid the import cycle, the repoinfo is moved
to a new package.
