+++ pkg/controller/master/upgrade/upgrade_controller.go
@@ -27,11 +27,13 @@ import (
 	kubevirtv1 "kubevirt.io/api/core/v1"
 
 	harvesterv1 "github.com/harvester/harvester/pkg/apis/harvesterhci.io/v1beta1"
+	"github.com/harvester/harvester/pkg/controller/master/upgrade/repoinfo"
 	ctlharvesterv1 "github.com/harvester/harvester/pkg/generated/controllers/harvesterhci.io/v1beta1"
 	kubevirtctrl "github.com/harvester/harvester/pkg/generated/controllers/kubevirt.io/v1"
 	ctllhv1 "github.com/harvester/harvester/pkg/generated/controllers/longhorn.io/v1beta2"
 	upgradectlv1 "github.com/harvester/harvester/pkg/generated/controllers/upgrade.cattle.io/v1"
 	"github.com/harvester/harvester/pkg/settings"
+	"github.com/harvester/harvester/pkg/upgradehelper/versionguard"
 	"github.com/harvester/harvester/pkg/util"
 )
 
@@ -276,7 +278,7 @@ func (h *upgradeHandler) OnChanged(_ string, upgrade *harvesterv1.Upgrade) (*har
 			Factor:   1.0,
 			Jitter:   0.1,
 		}
-		var repoInfo *RepoInfo
+		var repoInfo *repoinfo.RepoInfo
 		if err := retry.OnError(backoff, util.IsRetriableNetworkError, func() error {
 			repoInfo, err = repo.getInfo()
 			if err != nil {
@@ -294,9 +296,10 @@ func (h *upgradeHandler) OnChanged(_ string, upgrade *harvesterv1.Upgrade) (*har
 			setUpgradeCompletedCondition(toUpdate, StateFailed, corev1.ConditionFalse, err.Error(), "")
 			return h.upgradeClient.Update(toUpdate)
 		}
+		toUpdate.Status.RepoInfo = repoInfoStr
 
 		// Upgrade Eligibility Check
-		isEligible, reason := upgradeEligibilityCheck(upgrade, repoInfo)
+		isEligible, reason := upgradeEligibilityCheck(upgrade)
 
 		if !isEligible {
 			setUpgradeCompletedCondition(toUpdate, StateFailed, corev1.ConditionFalse, reason, "")
@@ -594,8 +597,8 @@ func ensureSingleUpgrade(namespace string, upgradeCache ctlharvesterv1.UpgradeCa
 	return onGoingUpgrades[0], nil
 }
 
-func getCachedRepoInfo(upgrade *harvesterv1.Upgrade) (*RepoInfo, error) {
-	repoInfo := &RepoInfo{}
+func getCachedRepoInfo(upgrade *harvesterv1.Upgrade) (*repoinfo.RepoInfo, error) {
+	repoInfo := &repoinfo.RepoInfo{}
 	if err := repoInfo.Load(upgrade.Status.RepoInfo); err != nil {
 		return nil, err
 	}
@@ -633,7 +636,7 @@ func isVersionUpgradable(currentVersion, minUpgradableVersion string) error {
 	return nil
 }
 
-func upgradeEligibilityCheck(upgrade *harvesterv1.Upgrade, repoInfo *RepoInfo) (bool, string) {
+func upgradeEligibilityCheck(upgrade *harvesterv1.Upgrade) (bool, string) {
 	skipVersionCheckStr, ok := upgrade.Annotations[skipVersionCheckAnnotation]
 	if ok {
 		skipVersionCheck, err := strconv.ParseBool(skipVersionCheckStr)
@@ -643,8 +646,7 @@ func upgradeEligibilityCheck(upgrade *harvesterv1.Upgrade, repoInfo *RepoInfo) (
 		}
 	}
 
-	logrus.Info("Check minimum upgradable version")
-	if err := isVersionUpgradable(upgrade.Status.PreviousVersion, repoInfo.Release.MinUpgradableVersion); err != nil {
+	if err := versionguard.Check(upgrade, true, ""); err != nil {
 		return false, err.Error()
 	}
