+++ pkg/upgradehelper/versionguard/repo.go
@@ -2,11 +2,11 @@ package versionguard
 
 import (
 	"github.com/harvester/harvester/pkg/apis/harvesterhci.io/v1beta1"
-	harvupgrade "github.com/harvester/harvester/pkg/controller/master/upgrade"
+	"github.com/harvester/harvester/pkg/controller/master/upgrade/repoinfo"
 )
 
-func getRepoInfo(upgrade *v1beta1.Upgrade) (*harvupgrade.RepoInfo, error) {
-	repoInfo := &harvupgrade.RepoInfo{}
+func getRepoInfo(upgrade *v1beta1.Upgrade) (*repoinfo.RepoInfo, error) {
+	repoInfo := &repoinfo.RepoInfo{}
 	if err := repoInfo.Load(upgrade.Status.RepoInfo); err != nil {
 		return nil, err
 	}
